+++ flutter_local_notifications/lib/src/platform_specifics/android/method_channel_mappers.dart
@@ -217,6 +217,7 @@ extension AndroidNotificationDetailsMapper on AndroidNotificationDetails {
         'additionalFlags': additionalFlags,
         'subText': subText,
         'tag': tag,
+        'colorized': colorized,
       }
         ..addAll(_convertStyleInformationToMap())
         ..addAll(_convertNotificationSoundToMap(sound))