+++ flutter_local_notifications/test/android_flutter_local_notifications_test.dart
@@ -2,6 +2,7 @@ import 'dart:typed_data';
 import 'dart:ui';
 import 'package:clock/clock.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_local_notifications/src/platform_specifics/android/enums.dart';
@@ -146,6 +147,7 @@ void main() {
                 'htmlFormatTitle': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -226,6 +228,7 @@ void main() {
                 'htmlFormatTitle': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -307,6 +310,7 @@ void main() {
                 'htmlFormatTitle': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -389,6 +393,7 @@ void main() {
                 'htmlFormatTitle': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -475,6 +480,7 @@ void main() {
                 'htmlFormatTitle': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -560,6 +566,7 @@ void main() {
                 'htmlFormatTitle': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -644,6 +651,7 @@ void main() {
                 'htmlFormatTitle': true,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -737,6 +745,7 @@ void main() {
                 'hideExpandedLargeIcon': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -840,6 +849,7 @@ void main() {
                 'hideExpandedLargeIcon': true,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -933,6 +943,7 @@ void main() {
                 'hideExpandedLargeIcon': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1036,6 +1047,7 @@ void main() {
                 'hideExpandedLargeIcon': true,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1126,6 +1138,7 @@ void main() {
                 'htmlFormatLines': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1223,6 +1236,7 @@ void main() {
                 'htmlFormatLines': true,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1305,6 +1319,7 @@ void main() {
                 'htmlFormatTitle': false,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1390,6 +1405,7 @@ void main() {
                 'htmlFormatTitle': true,
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1500,6 +1516,7 @@ void main() {
                 ],
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1623,6 +1640,7 @@ void main() {
                 ],
               },
               'tag': null,
+              'colorized': false,
             },
           }));
     });
@@ -1711,6 +1729,7 @@ void main() {
                       'htmlFormatTitle': false,
                     },
                     'tag': null,
+                    'colorized': false,
                   },
                 }));
           });
@@ -1804,6 +1823,7 @@ void main() {
                   'htmlFormatTitle': false,
                 },
                 'tag': null,
+                'colorized': false,
               },
             }));
       });
@@ -1896,6 +1916,7 @@ void main() {
                   'htmlFormatTitle': false,
                 },
                 'tag': null,
+                'colorized': false,
               },
             }));
       });
@@ -1989,6 +2010,7 @@ void main() {
                   'htmlFormatTitle': false,
                 },
                 'tag': null,
+                'colorized': false,
               },
             }));
       });
@@ -2198,6 +2220,106 @@ void main() {
           isMethodCall(
             'stopForegroundService',
             arguments: null,
+          ));
+    });
+
+    test('areNotificationsEnabled', () async {
+      await flutterLocalNotificationsPlugin
+          .resolvePlatformSpecificImplementation<
+              AndroidFlutterLocalNotificationsPlugin>()!
+          .areNotificationsEnabled();
+      expect(
+          log.last, isMethodCall('areNotificationsEnabled', arguments: null));
+    });
+
+    test('startForegroundServiceWithBlueBackgroundNotification', () async {
+      const AndroidInitializationSettings androidInitializationSettings =
+          AndroidInitializationSettings('app_icon');
+      const InitializationSettings initializationSettings =
+          InitializationSettings(android: androidInitializationSettings);
+      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
+      const AndroidNotificationDetails androidPlatformChannelSpecifics =
+          AndroidNotificationDetails(
+        'channelId',
+        'channelName',
+        channelDescription: 'channelDescription',
+        color: Colors.blue,
+        colorized: true,
+      );
+
+      await flutterLocalNotificationsPlugin
+          .resolvePlatformSpecificImplementation<
+              AndroidFlutterLocalNotificationsPlugin>()!
+          .startForegroundService(1, 'colored background notification title',
+              'colored background notification body',
+              notificationDetails: androidPlatformChannelSpecifics);
+      expect(
+          log.last,
+          isMethodCall(
+            'startForegroundService',
+            arguments: <String, Object?>{
+              'notificationData': {
+                'id': 1,
+                'title': 'colored background notification title',
+                'body': 'colored background notification body',
+                'payload': '',
+                'platformSpecifics': <String, Object?>{
+                  'icon': null,
+                  'channelId': 'channelId',
+                  'channelName': 'channelName',
+                  'channelDescription': 'channelDescription',
+                  'channelShowBadge': true,
+                  'channelAction':
+                      AndroidNotificationChannelAction.createIfNotExists.index,
+                  'importance': Importance.defaultImportance.value,
+                  'priority': Priority.defaultPriority.value,
+                  'playSound': true,
+                  'enableVibration': true,
+                  'vibrationPattern': null,
+                  'groupKey': null,
+                  'setAsGroupSummary': false,
+                  'groupAlertBehavior': GroupAlertBehavior.all.index,
+                  'autoCancel': true,
+                  'ongoing': false,
+                  'colorAlpha': 255,
+                  'colorRed': 33,
+                  'colorGreen': 150,
+                  'colorBlue': 243,
+                  'onlyAlertOnce': false,
+                  'showWhen': true,
+                  'when': null,
+                  'usesChronometer': false,
+                  'showProgress': false,
+                  'maxProgress': 0,
+                  'progress': 0,
+                  'indeterminate': false,
+                  'enableLights': false,
+                  'ledColorAlpha': null,
+                  'ledColorRed': null,
+                  'ledColorGreen': null,
+                  'ledColorBlue': null,
+                  'ledOnMs': null,
+                  'ledOffMs': null,
+                  'ticker': null,
+                  'visibility': null,
+                  'timeoutAfter': null,
+                  'category': null,
+                  'additionalFlags': null,
+                  'fullScreenIntent': false,
+                  'shortcutId': null,
+                  'subText': null,
+                  'style': AndroidNotificationStyle.defaultStyle.index,
+                  'styleInformation': <String, Object>{
+                    'htmlFormatContent': false,
+                    'htmlFormatTitle': false,
+                  },
+                  'tag': null,
+                  'colorized': true,
+                },
+              },
+              'startType': AndroidServiceStartType.startSticky.value,
+              'foregroundServiceTypes': null
+            },
           ));
     });
   });