+++ flutter_local_notifications_linux/CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.4.2]
+
+* Bumped dependencies. Thanks to PR from [Guy Luz](https://github.com/guyluz11)
+
+## [0.4.1+1]
+
+* Fixed minor casing error in 0.4.1 changelog entry
+
+## [0.4.1]
+
+* Fix `initialize()` returning null all the time instead of returning an appropriate boolean value to indicate if plugin has been initialised
+
+## [0.4.0]
+
+*  Bumped `dbus` dependency.
+
 ## [0.3.0]
 
 * **Breaking change** the `SelectNotificationCallback` typedef now maps to a function that returns `void` instead of a `Future<dynamic>`. This change was done to better communicate the plugin doesn't actually await any asynchronous computation and is similar to how button pressed callbacks work for Flutter where they are typically use [`VoidCallback`](https://api.flutter.dev/flutter/dart-ui/VoidCallback.html)